Add mined blocks count to context
