+++ testnet/stacks-node/src/tests/signer/commands/mod.rs
@@ -10,8 +10,8 @@ mod sortition;
 mod stacks_mining;
 mod transfer;
 
-pub use bitcoin_mining::{MineBitcoinBlock, MineBitcoinBlockTenureChangeMiner1};
-pub use block_commit::{BuildNextBitcoinBlock, SubmitBlockCommitMiner1, SubmitBlockCommitMiner2};
+pub use bitcoin_mining::{MineBitcoinBlock, MineBitcoinBlockTenureChangeMiner1, BuildNextBitcoinBlocks};
+pub use block_commit::{SubmitBlockCommitMiner1, SubmitBlockCommitMiner2};
 pub use block_wait::{
     VerifyMiner1BlockCount, WaitForAndVerifyBlockRejection, WaitForTenureChangeBlockFromMiner1,
     WaitForTenureChangeBlockFromMiner2,
